Moved current renderer to html directory
[#178712993]
