--- lib/prosemirror-to-html.rb
@@ -1,2 +1,0 @@
-require_relative 'prosemirror_to_html/renderer'
-require_relative 'prosemirror_to_html/version'
