+++ lib/prosemirror.rb
@@ -1,0 +1,3 @@
+require_relative 'prosemirror/html/renderer'
+require_relative 'prosemirror/text/renderer'
+require_relative 'prosemirror/version'
