+++ lib/prosemirror/html/marks/bold.rb
@@ -1,0 +1,15 @@
+require_relative './mark'
+
+module Prosemirror
+  module Html
+    module Marks
+      class Bold < Mark
+        def tags
+          [
+            { tag: 'strong' }
+          ]
+        end
+      end
+    end
+  end
+end
