+++ lib/prosemirror/html/marks/code.rb
@@ -1,0 +1,15 @@
+require_relative './mark'
+
+module Prosemirror
+  module Html
+    module Marks
+      class Code < Mark
+        def tags
+          [
+            { tag: 'code' }
+          ]
+        end
+      end
+    end
+  end
+end
