+++ lib/prosemirror/html/marks/italic.rb
@@ -1,0 +1,15 @@
+require_relative './mark'
+
+module Prosemirror
+  module Html
+    module Marks
+      class Italic < Mark
+        def tags
+          [
+            { tag: 'em' }
+          ]
+        end
+      end
+    end
+  end
+end
