+++ lib/prosemirror/html/marks/mark.rb
@@ -1,0 +1,13 @@
+module Prosemirror
+  module Html
+    module Marks
+      class Mark
+        attr_reader :node
+
+        def initialize(node)
+          @node = node
+        end
+      end
+    end
+  end
+end
