+++ lib/prosemirror/html/nodes/bullet_list.rb
@@ -1,0 +1,15 @@
+require_relative './node'
+
+module Prosemirror
+  module Html
+    module Nodes
+      class BulletList < Node
+        def tags
+          [
+            { tag: 'ul' }
+          ]
+        end
+      end
+    end
+  end
+end
