+++ lib/prosemirror/html/nodes/code_block.rb
@@ -1,0 +1,20 @@
+require_relative './node'
+
+module Prosemirror
+  module Html
+    module Nodes
+      class CodeBlock < Node
+        def tags
+          [
+            {
+              tag: 'pre'
+            },
+            {
+              tag: 'code'
+            }
+          ]
+        end
+      end
+    end
+  end
+end
