+++ lib/prosemirror/html/nodes/hard_break.rb
@@ -1,0 +1,15 @@
+require_relative './node'
+
+module Prosemirror
+  module Html
+    module Nodes
+      class HardBreak < Node
+        def tags
+          [
+            { tag: 'br' }
+          ]
+        end
+      end
+    end
+  end
+end
