+++ lib/prosemirror/html/nodes/list_item.rb
@@ -1,0 +1,15 @@
+require_relative './node'
+
+module Prosemirror
+  module Html
+    module Nodes
+      class ListItem < Node
+        def tags
+          [
+            { tag: 'li' }
+          ]
+        end
+      end
+    end
+  end
+end
