+++ lib/prosemirror/html/nodes/ordered_list.rb
@@ -1,0 +1,15 @@
+require_relative './node'
+
+module Prosemirror
+  module Html
+    module Nodes
+      class OrderedList < Node
+        def tags
+          [
+            { tag: 'ol' }
+          ]
+        end
+      end
+    end
+  end
+end
