+++ lib/prosemirror/html/nodes/text.rb
@@ -1,0 +1,10 @@
+require_relative './node'
+
+module Prosemirror
+  module Html
+    module Nodes
+      class Text < Node
+      end
+    end
+  end
+end
