+++ lib/prosemirror/version.rb
@@ -1,0 +1,3 @@
+module Prosemirror
+  VERSION = '0.2.0'.freeze
+end
