--- lib/prosemirror_to_html/marks/bold.rb
@@ -1,13 +1,0 @@
-require_relative './mark'
-
-module ProsemirrorToHtml
-  module Marks
-    class Bold < Mark
-      def tags
-        [
-          { tag: 'strong' }
-        ]
-      end
-    end
-  end
-end
