--- lib/prosemirror_to_html/marks/code.rb
@@ -1,13 +1,0 @@
-require_relative './mark'
-
-module ProsemirrorToHtml
-  module Marks
-    class Code < Mark
-      def tags
-        [
-          { tag: 'code' }
-        ]
-      end
-    end
-  end
-end
