--- lib/prosemirror_to_html/marks/italic.rb
@@ -1,13 +1,0 @@
-require_relative './mark'
-
-module ProsemirrorToHtml
-  module Marks
-    class Italic < Mark
-      def tags
-        [
-          { tag: 'em' }
-        ]
-      end
-    end
-  end
-end
