--- lib/prosemirror_to_html/marks/link.rb
@@ -1,13 +1,0 @@
-require_relative './mark'
-
-module ProsemirrorToHtml
-  module Marks
-    class Link < Mark
-      def tags
-        [
-          { tag: 'a', attrs: @node.attrs }
-        ]
-      end
-    end
-  end
-end
