--- lib/prosemirror_to_html/marks/mark.rb
@@ -1,11 +1,0 @@
-module ProsemirrorToHtml
-  module Marks
-    class Mark
-      attr_reader :node
-
-      def initialize(node)
-        @node = node
-      end
-    end
-  end
-end
