--- lib/prosemirror_to_html/marks/strike.rb
@@ -1,13 +1,0 @@
-require_relative './mark'
-
-module ProsemirrorToHtml
-  module Marks
-    class Strike < Mark
-      def tags
-        [
-          { tag: 's' }
-        ]
-      end
-    end
-  end
-end
