--- lib/prosemirror_to_html/nodes/bullet_list.rb
@@ -1,13 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class BulletList < Node
-      def tags
-        [
-          { tag: 'ul' }
-        ]
-      end
-    end
-  end
-end
