--- lib/prosemirror_to_html/nodes/code_block.rb
@@ -1,18 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class CodeBlock < Node
-      def tags
-        [
-          {
-            tag: 'pre'
-          },
-          {
-            tag: 'code'
-          }
-        ]
-      end
-    end
-  end
-end
