--- lib/prosemirror_to_html/nodes/hard_break.rb
@@ -1,13 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class HardBreak < Node
-      def tags
-        [
-          { tag: 'br' }
-        ]
-      end
-    end
-  end
-end
