--- lib/prosemirror_to_html/nodes/heading.rb
@@ -1,13 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class Heading < Node
-      def tags
-        [
-          { tag: "h#{data.attrs.level}" }
-        ]
-      end
-    end
-  end
-end
