--- lib/prosemirror_to_html/nodes/image.rb
@@ -1,13 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class Image < Node
-      def tags
-        [
-          { tag: 'img', attrs: data.attrs }
-        ]
-      end
-    end
-  end
-end
