--- lib/prosemirror_to_html/nodes/list_item.rb
@@ -1,13 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class ListItem < Node
-      def tags
-        [
-          { tag: 'li' }
-        ]
-      end
-    end
-  end
-end
