--- lib/prosemirror_to_html/nodes/loom_embed.rb
@@ -1,25 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class LoomEmbed < Node
-      def tags
-        [
-          {
-            tag: 'a',
-            attrs: {
-              href: data.attrs[:share]
-            }
-          },
-          {
-            tag: 'img',
-            attrs: {
-              src: data.attrs[:thumb],
-              style: 'max-width: 470px;'
-            }
-          }
-        ]
-      end
-    end
-  end
-end
