--- lib/prosemirror_to_html/nodes/mention.rb
@@ -1,17 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class Mention < Node
-      def tags
-        [
-          { tag: 'strong' }
-        ]
-      end
-
-      def text
-        data.attrs.label
-      end
-    end
-  end
-end
