--- lib/prosemirror_to_html/nodes/ordered_list.rb
@@ -1,13 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class OrderedList < Node
-      def tags
-        [
-          { tag: 'ol' }
-        ]
-      end
-    end
-  end
-end
