--- lib/prosemirror_to_html/nodes/paragraph.rb
@@ -1,13 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class Paragraph < Node
-      def tags
-        [
-          { tag: 'p' }
-        ]
-      end
-    end
-  end
-end
