--- lib/prosemirror_to_html/nodes/text.rb
@@ -1,8 +1,0 @@
-require_relative './node'
-
-module ProsemirrorToHtml
-  module Nodes
-    class Text < Node
-    end
-  end
-end
