--- lib/prosemirror_to_html/version.rb
@@ -1,3 +1,0 @@
-module ProsemirrorToHtml
-  VERSION = '0.1.0'.freeze
-end
